project/consumer/requirements.txt
fastapi
uvicorn
celery
redis
# flower
requests
numpy
opencv-python
shapely
lapx==0.5.9
ultralytics
python-dotenv
